Agregado de api. Fetch

+++ desafio2-1.js
@@ -28,10 +28,13 @@ function iniciarSesion()
             title: `Iniciaste sesion ${USUADMIN}`,
             icon: 'success',
             focusConfirm: true,
-            confirmButtonText: `<a href="./pages/facturas.html"><i class="fa fa-thumbs-up"></i><strong>CLICKEAME</strong></a>`
-        })
-
-
+            confirmButtonText: `<i class="fa fa-thumbs-up"></i><strong>EXITO!</strong>`
+        }
+        
+        )
+        setTimeout(()=>{
+            window.location.assign(`./pages/facturas.html`);
+        }, 2000)
     }
     else{
         // card.innerHTML="";
@@ -84,10 +87,21 @@ function agregarFactura()
             localStorage.setItem(USUADMIN, JSON.stringify([Factura]));
             verFacturas([Factura]);
 
-            card.classList.add(`card`);
-            card.innerHTML = `<p>Factura agregada</p>`;
-            document.body.append(card);
-            console.log("primer if anidado")
+            // card.classList.add(`card`);
+            // card.innerHTML = `<p>Factura agregada</p>`;
+            // document.body.append(card);
+            // console.log("primer if anidado")
+            
+            Swal.fire({
+                title: `Factura agregada`,
+                icon: 'success',
+                focusConfirm: true,
+                confirmButtonText: `<i class="fa fa-thumbs-up"></i><strong>EXITO!</strong>`,
+                timer: 1200
+            }
+            
+            )
+
             precioFacturas();
         }
     
@@ -102,37 +116,66 @@ function agregarFactura()
             localStorageFacturas.push(Factura);
             localStorage.setItem(USUADMIN, JSON.stringify(localStorageFacturas));
             verFacturas(localStorageFacturas);
-            card.classList.add(`card`);
-            card.innerHTML = `<p>Factura agregada</p>`;
-            document.body.append(card);
+            // card.classList.add(`card`);
+            // card.innerHTML = `<p>Factura agregada</p>`;
+            // document.body.append(card);
+            Swal.fire({
+                title: `Factura agregada`,
+                icon: 'success',
+                focusConfirm: true,
+                confirmButtonText: `<i class="fa fa-thumbs-up"></i><strong>EXITO!</strong>`,
+                timer: 1200
+            }
+            
+            )
             precioFacturas();
         }
         else if (ENCONTRADO!==undefined){
             console.log("ACA ESTOY");
-            card.innerHTML="";
-            card.classList.add(`card`);
-            card.innerHTML = `<p>La factura n° <strong>${inputNumeroFactura.value}</strong> ya fue cargada</p>`;
-            document.body.append(card);
+            // card.innerHTML="";
+            // card.classList.add(`card`);
+            // card.innerHTML = `<p>La factura n° <strong>${inputNumeroFactura.value}</strong> ya fue cargada</p>`;
+            // document.body.append(card);
+            Swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: `La factura n° ${inputNumeroFactura.value} ya fue cargada`
+              })
         }
         else if(inputCuit.value.length <11){
-            card.innerHTML="";
-            card.classList.add(`card`);
-            card.innerHTML = `<p>El cuit debe tener 11 digitos </p>`;
-            document.body.append(card);
+            // card.innerHTML="";
+            // card.classList.add(`card`);
+            // card.innerHTML = `<p>El cuit debe tener 11 digitos </p>`;
+            // document.body.append(card);
+            Swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: `El CUIT debe tener 11 digitos`
+              }) 
         }
         else if(isNaN(inputNumeroComprobante.value) !== false){
-            card.innerHTML="";
-            card.classList.add(`card`);
-            card.innerHTML = `<p>Ingrese un valor válido</p>`;
-            document.body.append(card);
+            // card.innerHTML="";
+            // card.classList.add(`card`);
+            // card.innerHTML = `<p>Ingrese un valor válido</p>`;
+            // document.body.append(card);
+            Swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: `Ingrese un valor válido`
+              })
         }
     }
     }
     else{
-                    card.innerHTML="";
-            card.classList.add(`card`);
-            card.innerHTML = `<p>Ingrese un valor válido</p>`;
-            document.body.append(card);
+            //         card.innerHTML="";
+            // card.classList.add(`card`);
+            // card.innerHTML = `<p>Ingrese un valor válido</p>`;
+            // document.body.append(card);
+            Swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: `Ingrese un valor válido`
+              })
     }
 
 }
@@ -144,7 +187,7 @@ function precioFacturas(){
     console.log(suma);
     let totalSumaFacturas = document.getElementById(`totalFacturas`);
 
-    totalSumaFacturas.innerHTML=`<p>$${suma}</p>`;   
+    totalSumaFacturas.innerHTML=`<p>TOTAL: $${suma}</p>`;   
     return (suma);
 }
 
@@ -162,12 +205,32 @@ function precioFacturas(){
 // }
 
 function eliminarFacturas(){
-    localStorage.clear();
-    listaFacturas.innerHTML=`<h1>No hay facturas cargadas</h1>`;
-    card.classList.add(`card`);
-    card.innerHTML = `<p>Eliminaste todas las facturas</p>`;
-    document.body.append(card);
-    totalFacturas.innerHTML = `<p>$0</p>`;
+    Swal.fire({
+        title: 'Estás seguro?',
+        text: "No podrás deshacer los cambios",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si! Borrar'
+      }).then((result) => {
+        if (result.isConfirmed) {
+            localStorage.clear();
+          Swal.fire(
+            'Facturas eliminadas!',
+            'Las facturas han sido eliminadas',
+            'Exito!'
+          )
+          listaFacturas.innerHTML=`<h1>No hay facturas cargadas</h1>`;
+          totalFacturas.innerHTML = `<p>$0</p>`;
+        }
+      })
+    // localStorage.clear();
+    // listaFacturas.innerHTML=`<h1>No hay facturas cargadas</h1>`;
+    // card.classList.add(`card`);
+    // card.innerHTML = `<p>Eliminaste todas las facturas</p>`;
+    // document.body.append(card);
+    // totalFacturas.innerHTML = `<p>$0</p>`;
 }
 
 btnEliminar.addEventListener(`click`,()=>{
@@ -182,4 +245,10 @@ function verFacturas(facturas){
         li.innerHTML = `<hr> Factura N°: ${factura.facturaN} - N° Comprobante: ${factura.comprobanteN} - CUIT: ${factura.cuit} - Dirección: ${factura.edificio} - Administración: ${factura.administracion} - Precio: $${factura.precio} - Descripción: ${factura.descripcion} - Fecha: ${factura.fecha}`;
         listaFacturas.appendChild(li);
     });
+}
+
+function verConsorcios(){
+fetch('data.json')
+.then(info=>info.json())
+.then(data=>console.log(data));
 }
